Read the notes; big change
Updated colors. This is better. The old one looked too 1999.
The buttons and gotopage now go to the next item on whatever you were
looking at in the text field. If you're stuck on notes, it moves you to
the next note.
Cleaned up the indentation of the document a bit.